requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
parse==1.15.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
requests==2.23.0
soupsieve==2.0.1
urllib3==1.25.9
Werkzeug==1.0.1
